Rename MS -> SA

+++ SCADA Import/PLC.cs
@@ -33,7 +33,7 @@ namespace JLR.SCADA.DCP
             Sequnces.Add(s.Key, s);
         }
 
-        public Sequence AddSequence(int seqNum, string zone, int ms, string description)
+        public Sequence AddSequence(int seqNum, string zone, double ms, string description)
         {
             Sequence s = new Sequence(this, seqNum, zone, ms, description);
             AddSequence(s);
